Deferred cleanups: CI format gate, helper relocation, site param (#396, #398, #408a)
Three small deferred-work items from a review of the last 50 PRs:

- #396: add a `ruff format --check` step to the Ruff workflow. Only
  `ruff check` ran, so formatting drift slipped through until a later
  commit reformatted it; this fails the build on any unformatted file,
  matching the pre-commit `ruff format` contributors run locally.

- #398: move the pure DPLA-provenance helpers `_is_dpla_reference` and
  `_entity_has_dpla_attributed_claims` from tools/sdc_sync.py into the
  shared ingest_wikimedia/sdc.py library and re-import them. Now colocated
  with the `Q_DPLA` constant, `_is_dpla_reference` uses it instead of the
  bare "Q2944483" literal.

- #408a: replace the uploader's cross-module `sdc_sync.site = self.site`
  mutation with a `commons_site=` parameter on `merge_item_onto_canonical`
  that installs the module global internally. The caller now passes site
  as data; the affected test asserts the parameter contract.

No behavior change intended. Full pytest suite (1301) green; ruff check and
format clean.
