+++ .github/workflows/ruff.yml
@@ -15,3 +15,10 @@ jobs:
           # Don't leave a writable GITHUB_TOKEN in the runner for later steps.
           persist-credentials: false
       - uses: astral-sh/ruff-action@278981a28ce3188b1e39527901f38254bf3aac89 # v4.1.0
+      # Also enforce formatting — `ruff check` (above) does not check format, so
+      # style drift slipped through until a later commit reformatted it. This
+      # step fails the build on any unformatted file (matches the pre-commit
+      # `ruff format` contributors run locally).
+      - uses: astral-sh/ruff-action@278981a28ce3188b1e39527901f38254bf3aac89 # v4.1.0
+        with:
+          args: format --check
