+++ ingest_wikimedia/sdc.py
@@ -2000,3 +2000,49 @@ def build_claims_for_doc(
         )
 
     return {"claims": claims, "ingest_date": retrieval_date.isoformat()}
+
+
+def _is_dpla_reference(reference) -> bool:
+    """Return True iff `reference` is a DPLA-authored reference, identified
+    by a `P123 = Q2944483` snak (publisher = "Digital Public Library of
+    America"). DPLA stamps that snak on every reference it writes via
+    formattedclaim, so it's a sufficient marker for "we authored this".
+    """
+    snaks = (reference or {}).get("snaks") or {}
+    for snak in snaks.get("P123") or []:
+        try:
+            if snak["datavalue"]["value"]["id"] == Q_DPLA:
+                return True
+        except (KeyError, TypeError):
+            continue
+    return False
+
+
+def _entity_has_dpla_attributed_claims(entity: dict) -> bool:
+    """True iff ``entity`` carries at least one DPLA-attributed statement.
+
+    Mirrors :func:`Module:DPLA`'s ``isDplaDetermined`` filter on Commons:
+    the canonical DPLA-provenance marker is the ``P123 = Q2944483``
+    (publisher = DPLA) snak in any statement reference, stamped by
+    ``formattedclaim`` on every claim DPLA writes. The prior implementation
+    checked the ``P459 = Q61848113`` qualifier instead — a display-side
+    co-stamp that isn't a reference-independent marker — despite the
+    docstring already claiming alignment with Module:DPLA. Detecting
+    provenance via the reference closes that drift and keeps this guard
+    aligned with the template renderer's notion of "has DPLA SDC".
+    """
+    if not entity:
+        return False
+    statements = entity.get("statements") or entity.get("claims") or {}
+    if not isinstance(statements, dict):
+        return False
+    for stmt_list in statements.values():
+        if not isinstance(stmt_list, list):
+            continue
+        for stmt in stmt_list:
+            if not isinstance(stmt, dict):
+                continue
+            for reference in stmt.get("references") or []:
+                if _is_dpla_reference(reference):
+                    return True
+    return False
